refactor(frontend): Set up React environment, added pages, redux, react touter, home page, dashboard
