+++ frontend/package.json
@@ -13,11 +13,15 @@
     "@fortawesome/fontawesome-free": "^6.6.0",
     "@heroicons/react": "^2.1.5",
     "@material-tailwind/react": "2.1.8",
+    "@mui/icons-material": "^6.1.7",
     "prop-types": "^15.8.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.28.0",
     "react-toastify": "^10.0.5",
-    "react-youtube": "^10.1.0"
+    "react-youtube": "^10.1.0",
+    "redux": "^5.0.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.15.0",
